# farming/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/file": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查詢全部照片跟影片",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "查询错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "上傳照片/影片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "傳送人",
                        "name": "Sender",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上傳檔案",
                        "name": "Image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "機器編號",
                        "name": "Machine_Id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "位置",
                        "name": "Location",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "上傳檔案錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/file/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "修改 {id} 上傳檔案",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "上傳檔案ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "傳送人",
                        "name": "Sender",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上傳檔案",
                        "name": "Image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "機器編號",
                        "name": "Machine_Id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "位置",
                        "name": "Location",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "修改上傳檔案錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/message": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询全部留言",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "查询错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新增留言",
                "parameters": [
                    {
                        "type": "string",
                        "description": "留言內容",
                        "name": "Content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "新增留言錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/message/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询 {id} 留言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "留言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "找不到留言",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新留言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "留言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "留言內容",
                        "name": "Content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "更新留言錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "刪除留言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "留言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "找不到留言",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}